feat: Make entity state, and entity about entityregistry
Why:

 * noticed that entity-registry has additional info (like device and platform)
   about an entity
 * to introduce that the pattern of <cmd> <op> would collapse as we early
   on mistakenly called state lookups for entity lookups.

This change addreses the need by:

 * move current entity to a state command
 * introduce new entity command that for now just can list from entity-registry
   and rename entity id and names.
 * remove redundant tests in test_event.py
 * reorg/rename the tests using `entity` to use `state` instead.

Fixes #189
