+++ homeassistant_cli/hassconst.py
@@ -462,3 +462,7 @@ WS_TYPE_AREA_REGISTRY_CREATE = 'config/area_registry/create'
 WS_TYPE_AREA_REGISTRY_DELETE = 'config/area_registry/delete'
 WS_TYPE_AREA_REGISTRY_UPDATE = 'config/area_registry/update'
 WS_TYPE_DEVICE_REGISTRY_UPDATE = 'config/device_registry/update'
+WS_TYPE_ENTITY_REGISTRY_LIST = 'config/entity_registry/list'
+WS_TYPE_ENTITY_REGISTRY_GET = 'config/entity_registry/get'
+WS_TYPE_ENTITY_REGISTRY_UPDATE = 'config/entity_registry/update'
+WS_TYPE_ENTITY_REGISTRY_REGISTRY = 'config/entity_registry/remove'
