+++ tests/test_plugins.py
@@ -9,6 +9,7 @@ DFEAULT_PLUGINS = [
     'completion',
     'config',
     'discover',
+    'state',
     'entity',
     'event',
     'info',
